Refactor Priority from Tag to Task package
