+++ src/main/java/seedu/address/model/task/Task.java
@@ -8,7 +8,6 @@ import java.util.Objects;
 import java.util.Set;
 
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.tag.Priority;
 import seedu.address.model.tag.Tag;
 
 /**
