+++ src/test/java/seedu/address/storage/JsonAdaptedTaskTest.java
@@ -11,7 +11,7 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.tag.Priority;
+import seedu.address.model.task.Priority;
 import seedu.address.model.task.Deadline;
 import seedu.address.model.task.Description;
 import seedu.address.model.task.Note;
